[test] handle exit status in travis-test.sh
For #46, this commit should fail travis, since the previous one got
glide.lock problem

+++ scripts/travis-test.sh
@@ -30,7 +30,11 @@ glide -v
 echo "enable go vendor feature"
 export GO15VENDOREXPERIMENT=1
 
-overalls -project=github.com/dyweb/Ayi -debug
+if overalls -project=github.com/dyweb/Ayi -debug; then 
+    exit 1
+fi
+
+# check the exit status
 
 echo "recover the fixture"
 rm -r fixture
